# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Ephemeral VM Lab API
  version: 1.0.0
  description: >
    REST API for the orion-complex Ephemeral VM Lab.
    Manages disposable VM environments across libvirt (Linux) and
    Apple Virtualization.framework (macOS) providers.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        provider: { type: string }
        provider_subject: { type: string }
        email: { type: string }
        email_domain: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [user, admin] }
        disabled: { type: integer }
        created_at: { type: integer }
    Node:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        host_os: { type: string, enum: [linux, macos] }
        host_arch: { type: string }
        cpu_cores: { type: integer }
        memory_bytes: { type: integer }
        disk_bytes_total: { type: integer }
        max_cpu_utilization_ratio: { type: number }
        max_memory_utilization_ratio: { type: number }
        max_disk_utilization_ratio: { type: number }
        max_running_envs: { type: integer }
        online: { type: integer }
        last_heartbeat_at: { type: integer, nullable: true }
    Image:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        provider: { type: string, enum: [libvirt, macos, virtualization] }
        guest_os: { type: string }
        guest_arch: { type: string }
        base_image_id: { type: string, nullable: true }
        created_at: { type: integer }
...
    EnvironmentEvent:
      type: object
      properties:
        id: { type: string }
        env_id: { type: string }
        event_type: { type: string }
        from_state: { type: string, nullable: true }
        to_state: { type: string, nullable: true }
        triggered_by: { type: string, nullable: true }
        message: { type: string, nullable: true }
        created_at: { type: integer }
...
    NodeUsage:
      type: object
      properties:
        node_id: { type: string }
        environment_count: { type: integer }
        max_environments: { type: integer }
        vcpus_used: { type: integer }
        vcpus_capacity: { type: integer }
        memory_bytes_used: { type: integer }
        memory_bytes_capacity: { type: integer }
        disk_bytes_used: { type: integer }
        disk_bytes_capacity: { type: integer }
...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
security:
  - bearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      summary: Login via OIDC provider
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, access_token]
              properties:
                provider: { type: string, enum: [google, microsoft] }
                access_token: { type: string }
      responses:
        '200':
          description: Session token and user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
...
  /v1/nodes:
    get:
      summary: List nodes
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Node' }
    post:
      summary: Register node (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, host_os, host_arch, cpu_cores, memory_bytes, disk_bytes_total]
              properties:
                name: { type: string }
                host_os: { type: string }
                host_arch: { type: string }
                cpu_cores: { type: integer }
                memory_bytes: { type: integer }
                disk_bytes_total: { type: integer }
                max_cpu_utilization_ratio: { type: number }
                max_memory_utilization_ratio: { type: number }
                max_disk_utilization_ratio: { type: number }
                max_running_envs: { type: integer }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Node' }
  /v1/nodes/{node_id}:
    get:
      summary: Get node details
      parameters:
        - name: node_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Node' }
    put:
      summary: Update node (admin only)
      parameters:
        - name: node_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                max_cpu_utilization_ratio: { type: number }
                max_memory_utilization_ratio: { type: number }
                max_disk_utilization_ratio: { type: number }
                max_running_envs: { type: integer }
                online: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Node' }
    delete:
      summary: Delete node (admin only, must have no active environments)
      parameters:
        - name: node_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Node deleted
  /v1/nodes/{node_id}/heartbeat:
    post:
      summary: Node heartbeat
      parameters:
        - name: node_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Heartbeat recorded
  /v1/nodes/{node_id}/usage:
    get:
      summary: Get node resource utilization
      parameters:
        - name: node_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NodeUsage' }

  # ── Images ─────────────────────────────────────────────────────────
  /v1/images:
    get:
      summary: List images
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Image' }
    post:
      summary: Register image (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, provider, guest_os, guest_arch]
              properties:
                name: { type: string }
                provider: { type: string }
                guest_os: { type: string }
                guest_arch: { type: string }
                base_image_id: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Image' }
  /v1/images/{image_id}:
    get:
      summary: Get image details
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Image' }
    delete:
      summary: Delete image (admin only, blocked if in use or has children)
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Image deleted

  # ── Environments ───────────────────────────────────────────────────
...
  /v1/environments/{env_id}/events:
    get:
      summary: List environment events (audit trail)
      parameters:
        - name: env_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EnvironmentEvent' }